\name{IDTurtle-package}
\alias{IDTurtle-package}
\alias{IDTurtle}
\docType{package}
\title{
Individual identification of turtles by their plastral biometries
}
\description{
This package is very helpful to identify specimens of turtles with any marking, avoiding injuries or external identification devices on turtles.
This is also helpful to detect identification errors in databases
}
\details{
\tabular{ll}{
Package: \tab IDTurtle\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-23\cr
License: \tab GPL (version 2 or later)\cr
}
This package can be used for research, monitoring and conservation activities on turtles with plastral scutes.
}
\author{
Aitor Valdeon
Maintainer: Aitor Valdeon <emys@galapagosdenavarra.com>
}
\references{
Valdeon, A., Longares, L.A. (2015) IDTurtle: A new and safe method for individual turtle identification using R. (to complete once the article is accepted)
}
\keyword{ package }
\keyword{ turtle }
\keyword{ monitoring }
\keyword{ identification }
\keyword{ conservation }